pyproject.toml
[project]
name = "to-do-backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "django>=6.0.4",
    "django-cors-headers>=4.9.0",
    "djangorestframework>=3.17.1",
    "gunicorn>=25.3.0",
    "whitenoise>=6.12.0",
]
